requirements.txt
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1
fastapi~=0.111.0
psycopg2
uvicorn~=0.30.1
pydantic~=2.7.3
python-multipart
python-jose
passlib[bcrypt]